Create thread policies.

+++ app/Http/Controllers/ThreadsController.php
@@ -113,11 +113,14 @@ class ThreadsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param  \App\Channel $channel
      * @param  \App\Thread $thread
      * @return \Illuminate\Http\Response
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         if (request()->wantsJson()) {
